Implement `addQuizScoreByStudentId` endpoint.

+++ src/main/java/com/example/historian_api/dtos/requests/AddQuizScoreRequestDto.java
@@ -4,8 +4,8 @@ import java.math.BigDecimal;
 import java.util.List;
 
 public record AddQuizScoreRequestDto(
-        BigDecimal actualScore,
-        BigDecimal totalScore,
+        Integer actualScore,
+        Integer totalScore,
         BigDecimal time,
         List<QuestionAnswerRequestDto> questions
 ) {
